# Performance Tuning of Beneficiary Page Queries