This PR improved the speed of the beneficiaries page operations by using bulk inserts and adding a transactions block over the multiple update and delete statements:

- optimised queries to add and remove `tags` (adding tags query tested for 500 records and result speed improved from 3.2 seconds to 0.039 seconds using bulk inserts; removing tags query tested for 500 records and result speed improved from 2.15 seconds to 0.6 seconds using transaction) block and removing extra select)
- optimised query for the `touch` feature (Query speed checked for 500 records and speed enhanced from 6.2 seconds to 0.54 seconds using  transaction blocks over UPDATE statements and bulk inserts)
- added new method for bulk inserting changes history (`simpleBulkSaveChangeHistory`)
- added transaction block over multiple updates inside the `detached`, `merge`, `realdelete`,  `undelete` actions and also `correctdrops` method optimized by transaction block over update queries
- `bulkcorrectdrops` method created to apply the bulk inserts
- added `listBulkRealDelete` and `listBulkMove` to the list.php
- added `listBulkUndelete` and `listBulkUnDeleteAction` to the list.php