@vahidbazzaz 2021-12-29
![](https://github.trello.services/images/mini-trello-icon.png) [POST request for /?action=people takes ages](https://trello.com/c/87jwpyDV/795-post-request-for-actionpeople-takes-ages)